+++ app/Models/User.php
@@ -8,24 +8,32 @@ use Filament\Panel;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
+use Laravel\Sanctum\HasApiTokens;
 use Spatie\Permission\Traits\HasRoles;
 
 class User extends Authenticatable implements FilamentUser
 {
-    use HasFactory, Notifiable, HasRoles, HasPanelShield;
+    use HasFactory, Notifiable, HasRoles, HasPanelShield, HasApiTokens;
 
     protected $fillable = [
         'name',
         'email',
+        'avatar_url',
+        'is_active',
+        'identity_type',
         'password',
-        'must_change_password',
         'sipetra_id',
-        'sipetra_token',
-        'sipetra_refresh_token',
         'nip',
         'jabatan',
+        'must_change_password',
     ];
 
+    protected $casts = [
+        'is_active' => 'boolean',
+        'must_change_password' => 'boolean',
+        'email_verified_at' => 'datetime',
+        'password' => 'hashed',
+    ];
 
     protected $hidden = [
         'password',